pyproject.toml
[project]
name = "reddit-los-angeles-analysis"
version = "0.1.0"
description = "Analyze Reddit thread about moving to Los Angeles"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.18.0",
    "ipykernel>=6.29.5",
    "pandas>=2.2.3",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "tenacity>=9.1.2",
]
